# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Loja de Jogos API
  description: API para gerenciamento de jogos, usuários, pedidos, categorias e carrinho.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/pedidos:
    get:
      summary: Lista todos os pedidos
      responses:
        '200':
          description: Lista de pedidos
    post:
      summary: Cria um novo pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pedido'
      responses:
        '200':
          description: Pedido criado
  /api/pedidos/{id}:
    get:
      summary: Busca pedido por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido encontrado
        '404':
          description: Não encontrado
    put:
      summary: Atualiza um pedido
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pedido'
      responses:
        '200':
          description: Pedido atualizado
        '404':
          description: Não encontrado
    delete:
      summary: Remove um pedido
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Removido com sucesso
        '404':
          description: Não encontrado
...
components:
  schemas:
...
    Pedido:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dataCriacao:
          type: string
        itens:
          type: array
          items:
            type: object
...